# doc/api.yaml
openapi: 3.0.3
info:
  title: API WASAtext
  description: |-
    API necessarie a WASAText.
  version: "1.0.0"
...
paths:
...
  /wasachat/{utente}/nome:
    parameters:
    - name: utente
      in: path
      required: true
      description: Nickname dell'utente
      schema:
        type: string
        example: "Carlo"
        minLength: 1
        maxLength: 50
        pattern: '^[a-zA-Z0-9._~-]+$'
    put:
      tags: ["Profilo Utente"]
      summary: Modifica nome utente (setMyUserName)
      description: |-
        Modifica il nome utente.
      operationId: setMyUserName
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RichiestaNomeUtente'
        required: true
      responses:
        "200":
          description: Nome aggiornato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "202":
          description: Nome uguale al precedente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Utente non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
  /wasachat/{utente}/foto:
    parameters:
    - name: utente
      in: path
      required: true
      description: Nickname dell'utente
      schema:
        type: string
        example: "Carlo"
        minLength: 1
        maxLength: 50
        pattern: '^[a-zA-Z0-9._~-]+$'
    put:
      tags: ["Profilo Utente"]
      summary: Modifica foto profilo (setMyPhoto)
      description: |-
        Imposta o modifica la foto profilo dell'utente.
      operationId: setMyPhoto
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RichiestaFotoProfilo'
        required: true
      responses:
        "200":
          description: Foto aggiornata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaGenerica'
        "400":
          description: Errore lato client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Utente non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
...
  /wasachat/{utente}/utenti/gruppi/{gruppo}:     
    parameters:
    - name: utente
      in: path
      required: true
      description: Nickname dell'utente
      schema:
        type: string
        example: "Carlo"
        minLength: 1
        maxLength: 50
        pattern: '^[a-zA-Z0-9._~-]+$'

    - name: gruppo
      in: path
      required: true
      description: ID della chat
      schema:
        type: number
        example: 2
        minimum: 1
        maximum: 1024
        
    get:
      tags: ["Chat"]
      summary: get Utenti presenti nel gruppo (usersInGroup)
      description: |-
       Vedi chi sono gli utenti preesnti nel gruppo.
      operationId: usersInGroup
      responses:
        "200":
          description: Vengono restituiti correttamente gli utenti del gruppo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RispostaUtenti'
        "400":
          description: Errore lato client, l'utente non ha accesso al gruppo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "404":
          description: Messaggio non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
        "500":
          description: Errore interno del server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroreGenerico'
...
components:
  schemas:
...
    RichiestaNomeUtente:
      type: object
      description: Richiesta per modificare il nome utente
      required:
        - nome
      properties:
        nome:
          type: string
          description: nome
          example: "Peppe idraulico"
          minLength: 1
          maxLength: 100
          pattern: '^.*$'
    RichiestaFotoProfilo:
      type: object
      description: Richiesta per modificare la foto profilo
      required:
        - foto
      properties:
        foto:
          type: string
          description: Contenuto base64 della foto
          example: "fasfgasdkjfvashfadsfvhjfasfbashbvdfj--"
          minLength: 1
          maxLength: 1024
          pattern: '^.*$'
...
    RispostaUtenti:
      type: object
      description: Elenco dei utenti in un gruppo
      required:
        - utenti
      properties:
        utenti:
          type: array
          description: Lista degli utenti nel gruppo
          minItems: 0
          maxItems: 1024
          items:
            $ref: '#/components/schemas/Utenti'
    Utenti:
      type: object
      description: Utente, nome e immagine profilo
      required:
        - nome
        - foto
      properties:
        nome:
          type: string
          description: nome
          example: "Peppe idraulico"
          minLength: 1
          maxLength: 100
          pattern: '^.*$'
        foto:
          type: string
          description: Contenuto base64 della foto
          example: "fasfgasdkjfvashfadsfvhjfasfbashbvdfj--"
          minLength: 1
          maxLength: 1024
          pattern: '^.*$'
...
    RispostaGenerica:
      type: object
      description: Risposta generica
      required:
        - codice
        - risposta
      properties:
        codice:
          type: integer
          description: codice di risposta
          example: 200
          minLength: 1
          maxLength: 1024
          minimum: 200
          maximum: 299
        risposta:
          type: string
          description: risposta 
          example: "messaggio di successo"
          minLength: 1
          maxLength: 1024
          pattern: '^.*$'
    ErroreGenerico:
      type: object
      description: Errore
      required:
        - codice
        - errore
      properties:
        codiceErrore:
          type: integer
          description: codice di errore
          example: 500
          minLength: 1
          maxLength: 1024
          minimum: 300
          maximum: 599
        errore:
          type: string
          description: errore
          pattern: '^.*$'
          example: "errore"
          minLength: 1
          maxLength: 1024